feat: rescale probs and fix issues in computations (it works correctly now!)

+++ dlc/markov_chain.h
@@ -13,17 +13,17 @@ struct markov_chain {
     u32 num_states;
     u32 curr_state;
 
-    /* Вероятности scaled на 0..10000 (для точности 0.01%) */
-    u16 transition_probs[MC_MAX_STATES][MC_MAX_STATES];
+    /* Вероятности scaled на 0..DLC_PROB_SCALE (для точности 0.001%) */
+    u32 transition_probs[MC_MAX_STATES][MC_MAX_STATES];
 
-    /* начальное распределение состояний scaled на 0..10000 (0.01%) */
-    u16 init_distribution[MC_MAX_STATES];
+    /* начальное распределение состояний scaled на 0..DLC_PROB_SCALE */
+    u32 init_distribution[MC_MAX_STATES];
 };
 
 void markov_chain_init(struct markov_chain *mc, u32 num_states, 
                        struct dlc_state *states_array, 
-                       u16 transition_probs[][MC_MAX_STATES],
-                       u16 init_distribution[MC_MAX_STATES]);
+                       u32 transition_probs[][MC_MAX_STATES],
+                       u32 init_distribution[MC_MAX_STATES]);
 
 struct dlc_state* markov_chain_step(struct markov_chain *mc);
 
@@ -36,14 +36,14 @@ struct markov_chain_const {
     u32 num_states;
     u32 curr_state;
 
-    u16 transition_probs[MC_MAX_STATES][MC_MAX_STATES];
-    u16 init_distribution[MC_MAX_STATES];
+    u32 transition_probs[MC_MAX_STATES][MC_MAX_STATES];
+    u32 init_distribution[MC_MAX_STATES];
 };
 
 void markov_chain_const_init(struct markov_chain_const *mc, u32 num_states, 
     struct dlc_const_state *states_array, 
-    u16 transition_probs[][MC_MAX_STATES],
-    u16 init_distribution[MC_MAX_STATES]);
+    u32 transition_probs[][MC_MAX_STATES],
+    u32 init_distribution[MC_MAX_STATES]);
 
 struct dlc_const_state* markov_chain_const_step(struct markov_chain_const *mc);
